+++ src/app.rs
@@ -20,7 +20,7 @@ use crate::{
 		PushPopup, PushTagsPopup, RemoteListPopup, RenameBranchPopup,
 		RenameRemotePopup, ResetPopup, RevisionFilesPopup,
 		StashMsgPopup, SubmodulesListPopup, TagCommitPopup,
-		TagListPopup, UpdateRemoteUrlPopup,
+		TagListPopup, UpdateRemoteUrlPopup, WorktreesPopup,
 	},
 	queue::{
 		Action, AppTabs, InternalEvent, NeedsUpdate, Queue,
@@ -97,6 +97,7 @@ pub struct App {
 	select_branch_popup: BranchListPopup,
 	options_popup: OptionsPopup,
 	submodule_popup: SubmodulesListPopup,
+	worktree_popup: WorktreesPopup,
 	tags_popup: TagListPopup,
 	reset_popup: ResetPopup,
 	checkout_option_popup: CheckoutOptionPopup,
@@ -221,6 +222,7 @@ impl App {
 			tags_popup: TagListPopup::new(&env),
 			options_popup: OptionsPopup::new(&env),
 			submodule_popup: SubmodulesListPopup::new(&env),
+			worktree_popup: WorktreesPopup::new(&env),
 			log_search_popup: LogSearchPopupPopup::new(&env),
 			fuzzy_find_popup: FuzzyFindPopup::new(&env),
 			do_quit: QuitState::None,
@@ -523,6 +525,7 @@ impl App {
 			select_branch_popup,
 			revision_files_popup,
 			submodule_popup,
+			worktree_popup,
 			tags_popup,
 			options_popup,
 			help_popup,
@@ -552,6 +555,7 @@ impl App {
 			rename_remote_popup,
 			update_remote_url_popup,
 			submodule_popup,
+			worktree_popup,
 			tags_popup,
 			reset_popup,
 			checkout_option_popup,
@@ -796,6 +800,9 @@ impl App {
 			}
 			InternalEvent::ViewSubmodules => {
 				self.submodule_popup.open()?;
+			}
+			InternalEvent::ViewWorktrees => {
+				self.worktree_popup.open()?;
 			}
 			InternalEvent::Tags => {
 				self.tags_popup.open()?;