+++ src/queue.rs
@@ -147,6 +147,8 @@ pub enum InternalEvent {
 	///
 	ViewSubmodules,
 	///
+	ViewWorktrees,
+	///
 	ViewRemotes,
 	///
 	CreateRemote,